+++ source/Examples/Resources/CMClientPushSettings/CMClientPushSettings_Disabled.ps1
@@ -1,0 +1,13 @@
+Configuration Example
+{
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        CMClientPushSettings Clientsettings
+        {
+            SiteCode                              = 'Lab'
+            EnableAutomaticClientPushInstallation = $false
+        }
+    }
+}