+++ source/Examples/Resources/CMClientPushSettings/CMClientPushSettings_Include.ps1
@@ -1,0 +1,14 @@
+Configuration Example
+{
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        CMClientPushSettings Clientsettings
+        {
+            SiteCode          = 'Lab'
+            AccountsToInclude = @('contoso\Push1','contoso\Push2')
+            AccountsToExclude = @('contoso\Push3','contoso\Push4')
+        }
+    }
+}