+++ CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CMUserDiscovery Resource
 - Added CMSecurityRoles Resource
 - Added Add-DPToDPGroup to the ResourceHelper
+- Added CMClientPushSettings Resource
 
 ### Changed
 
@@ -78,6 +79,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   1910 and greater.
 - Updated ReadMe with latest versions of ConfigMgr that the module has been
   tested on.
+- Added DistributionPointInstallIis parameter to CmIniFile #62
+- Added an example for a Standalone Primary Site Server and updated required modules
+  to support.
 
 ### Removed